+++ packages/echo-core/src/actions/plantsData.ts
@@ -24,7 +24,20 @@ export function setPlantsData(partialPlantsData: Partial<PlantsData>): void {
  * @return {*}  {Readonly<PlantsData>}
  */
 export function getPlantsData(): Readonly<PlantsData> {
-    return readState(getCoreContext(), (state: GlobalState) => state.plantsData);
+    const plants = readState(getCoreContext(), (state: GlobalState) => state.plantsData);
+    if (plants.plants.length > 0) {
+        return plants;
+    }
+
+    const plantsFromStorage = persistPlantsData.getPlantsDataFromLocalStorage();
+    if (plantsFromStorage.plants.length > 0) {
+        dispatch(getCoreContext(), (state: GlobalState) => {
+            const newState = { ...state, plantsFromStorage };
+            return newState;
+        });
+    }
+
+    return plantsFromStorage;
 }
 
 /**